split namco21 driver into 3 drivers as the different configurations really are entirely different boardsets with similar components, not a real 'system' emulated entire PCB set for driveyes ( http://www.tvspels-nostalgi.com/Bilder/PCB/Namco/driverseye_cage_inside.jpg ) although how the PCBs communicate is not yet known (C139 maybe, which might also be an MCU)
