+++ src/mame/machine/namco_c139.cpp
@@ -8,6 +8,7 @@
     TODO:
     - Make this to actually work!
     - Is RAM shared with a specific CPU other than master/slave?
+	- is this another MCU with internal ROM?
 
 ***************************************************************************/
